added better check if server can receive signal in mandatory

+++ srcs/server.c
@@ -6,7 +6,7 @@
 /*   By: llevasse <llevasse@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/20 13:22:34 by llevasse          #+#    #+#             */
-/*   Updated: 2023/05/31 22:15:01 by llevasse         ###   ########.fr       */
+/*   Updated: 2023/06/13 14:55:23 by llevasse         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,10 +28,13 @@ void	sig_handler(int sig, siginfo_t *siginfo, void *context)
 			ft_lstadd_back(&g_sig_char.mini_str, ft_lstnew(g_sig_char.c));
 		if (g_sig_char.c == '\0')
 		{
-			ft_lstprint(g_sig_char.mini_str);
-			g_sig_char.mini_str = NULL;
-			if (kill(siginfo->si_pid, SIGUSR2) == -1)
-				ft_exit("Error in sending signal", 0);
+			if (g_sig_char.nb_null_received++ >= 1)
+			{
+				ft_lstprint(g_sig_char.mini_str);
+				g_sig_char.mini_str = NULL;
+				if (kill(siginfo->si_pid, SIGUSR2) == -1)
+					ft_exit("Error in sending signal", 0);
+			}
 		}
 		g_sig_char.shift = 7;
 		g_sig_char.c = 0;
@@ -59,6 +62,7 @@ int	main(void)
 	sa.sa_flags = SA_SIGINFO;
 	sigemptyset(&sa.sa_mask);
 	ft_printf("pid : %i\n", pid);
+	g_sig_char.nb_null_received = 0;
 	sigaction(SIGUSR1, &sa, NULL);
 	sigaction(SIGUSR2, &sa, NULL);
 	if (sigaction(SIGUSR2, &sa, NULL) < 0 || sigaction(SIGUSR1, &sa, NULL) < 0)
